+++ lib/ui/widgets/html_preview.dart
@@ -3,13 +3,14 @@ import 'dart:async';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:webview_flutter/webview_flutter.dart';
+import 'package:widgetbook/widgetbook.dart' show KnobsExtension;
 import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
 
-/// Wraps [WebViewWidget] to display HTML directly from [htmlDataSource].
+/// Wraps [WebViewWidget] to display HTML directly from [source].
 class HtmlPreview extends StatefulWidget {
-  final FutureOr<String> htmlDataSource;
+  final FutureOr<String> source;
 
-  const HtmlPreview({super.key, required this.htmlDataSource});
+  const HtmlPreview({super.key, required this.source});
 
   @override
   State<HtmlPreview> createState() => _HtmlPreviewState();
@@ -24,7 +25,7 @@ class _HtmlPreviewState extends State<HtmlPreview> {
 
     controller = WebViewController();
 
-    Future.value(widget.htmlDataSource).then(controller.loadHtmlString);
+    unawaited(_loadHtml());
   }
 
   @override
@@ -40,8 +41,8 @@ class _HtmlPreviewState extends State<HtmlPreview> {
   void didUpdateWidget(covariant HtmlPreview oldWidget) {
     super.didUpdateWidget(oldWidget);
 
-    if (oldWidget.htmlDataSource != widget.htmlDataSource) {
-      Future.value(widget.htmlDataSource).then(controller.loadHtmlString);
+    if (oldWidget.source != widget.source) {
+      unawaited(_loadHtml());
     }
   }
 
@@ -51,15 +52,29 @@ class _HtmlPreviewState extends State<HtmlPreview> {
       controller: controller,
     );
   }
+
+  /// Loads the HTML content from [widget.source] but also disables the
+  /// JavaScript before.
+  Future<void> _loadHtml() async {
+    final html = await widget.source;
+
+    controller.loadHtmlString(html);
+  }
 }
 
 @widgetbook.UseCase(
   path: '[Core]',
-  name: 'HtmlPreview',
+  name: '',
   type: HtmlPreview,
 )
 Widget previewHtmlPreview(BuildContext context) {
+  final html = context.knobs.string(
+    label: 'HTML',
+    initialValue: "<html lang=en><h1>Hello world!</h1></html>",
+    maxLines: 20,
+  );
+
   return HtmlPreview(
-    htmlDataSource: Future.value("<html lang=en><h1>Hello world!</h1></html>"),
+    source: Future.value(html),
   );
 }