+++ lib/ui/widgets/option_picker.dart
@@ -2,8 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:widgetbook_annotation/widgetbook_annotation.dart' as widgetbook;
 
 import '../../data/pdf_signing_option.dart';
-import '../app_theme.dart';
-import 'certificate_picker.dart';
+import 'radio_button.dart';
 import 'signature_type_picker.dart';
 
 /// Reusable widget to pick single option.
@@ -28,31 +27,28 @@ class OptionPicker<T> extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ListView.builder(
-      itemCount: values.length,
-      itemBuilder: (context, index) {
-        final value = values[index];
-        final child = _listItem(value);
+    return RadioGroup<T>(
+      groupValue: selectedValue,
+      onChanged: onValueChanged,
+      child: ListView.builder(
+        itemCount: values.length,
+        itemBuilder: (context, index) {
+          final value = values[index];
+          final child = _listItem(value);
 
-        return Padding(
-          padding: kMaterialListPadding,
-          child: child,
-        );
-      },
+          return Padding(
+            padding: kMaterialListPadding,
+            child: child,
+          );
+        },
+      ),
     );
   }
 
   Widget _listItem(T value) {
     final labelBuilder = this.labelBuilder ?? _defaultLabelBuilder;
     final label = Expanded(child: labelBuilder(value));
-    final radio = Transform.scale(
-      scale: kRadioScale,
-      child: Radio<T>(
-        value: value,
-        groupValue: selectedValue,
-        onChanged: onValueChanged,
-      ),
-    );
+    final radio = RadioButton<T>(value: value);
 
     return Semantics(
       checked: value == selectedValue,
@@ -65,9 +61,9 @@ class OptionPicker<T> extends StatelessWidget {
           },
           child: Row(
             mainAxisAlignment: MainAxisAlignment.start,
+            spacing: 8,
             children: [
               ExcludeSemantics(child: radio),
-              const SizedBox(width: 8),
               label,
             ],
           ),
@@ -97,12 +93,15 @@ Widget previewOptionPicker(BuildContext context) {
 
   return StatefulBuilder(
     builder: (context, setState) {
-      return OptionPicker(
-        values: PdfSigningOption.values,
-        selectedValue: selectedValue,
-        onValueChanged: (value) {
-          setState(() => selectedValue = value);
-        },
+      return Padding(
+        padding: const EdgeInsets.all(24),
+        child: OptionPicker(
+          values: PdfSigningOption.values,
+          selectedValue: selectedValue,
+          onValueChanged: (value) {
+            setState(() => selectedValue = value);
+          },
+        ),
       );
     },
   );